\name{ADAPTMSNM-package}
\alias{ADAPTMSNM}
\alias{ADAPTMSNM-package}
\title{ADAPTMSNM: ADAPT-MS Classifiers for Proteomics Data}
\description{
The \code{ADAPTMSNM} package provides R implementations of the ADAPT-MS
classifier family for proteomics workflows.
}
\details{
Primary user-facing objects are reference class generators:
\code{AdaptmsClassifierDF}, \code{AdaptmsClassifierFolder},
\code{AdaptmsMulticlassClassifier}, \code{BaselineClassifier},
\code{NonRefitClassifier}, \code{NonValClassifier},
\code{XGBRFClassifierDF}, and \code{XGBRFClassifierFolder}.
}
\docType{package}
